+++ src/main/java/com/example/springbootbackend/repository/PostRepository.java
@@ -3,5 +3,5 @@ package com.example.springbootbackend.repository;
 import com.example.springbootbackend.model.Post;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostRepository extends CrudRepository<Post, Integer> {
+public interface PostRepository extends CrudRepository<Post, Long> {
 }
